Refactor notification init and policy normalization
