+++ worker/services/NotificationPublisher.ts
@@ -1,4 +1,12 @@
-import type { Env, NotificationQueueMessage, NotificationRecipientSnapshot, NotificationCategory } from '../types.js';
+import type {
+  Env,
+  NotificationQueueMessage,
+  NotificationRecipientSnapshot,
+  NotificationCategory,
+  NotificationPolicy,
+  NotificationPolicyCategoryKey
+} from '../types.js';
+import { normalizeNotificationPolicy } from '../types.js';
 import { RemoteApiService } from './RemoteApiService.js';
 import { Logger } from '../utils/logger.js';
 
@@ -11,17 +19,6 @@ interface PracticeMember {
 const ADMIN_ROLES = new Set(['owner', 'admin']);
 
 type RemoteNotificationPreferences = Record<string, unknown>;
-type NotificationPolicyCategoryKey = 'messages' | 'system' | 'payments' | 'intakes' | 'matters';
-
-interface NotificationPolicyChannel {
-  push: boolean;
-  email: boolean;
-}
-
-interface NotificationPolicy {
-  defaults: Record<NotificationPolicyCategoryKey, NotificationPolicyChannel>;
-  allowed: Record<NotificationPolicyCategoryKey, NotificationPolicyChannel>;
-}
 
 const categoryPreferenceKey: Record<NotificationCategory, { push: string; email: string }> = {
   message: { push: 'messages_push', email: 'messages_email' },
@@ -38,75 +35,6 @@ const categoryPolicyKey: Record<NotificationCategory, NotificationPolicyCategory
   matter: 'matters',
   system: 'system'
 };
-
-const DEFAULT_ALLOWED_POLICY: Record<NotificationPolicyCategoryKey, NotificationPolicyChannel> = {
-  messages: { push: true, email: true },
-  system: { push: true, email: true },
-  payments: { push: true, email: true },
-  intakes: { push: true, email: true },
-  matters: { push: true, email: true }
-};
-
-const DEFAULT_NOTIFICATION_POLICY: NotificationPolicy = {
-  defaults: {
-    messages: { ...DEFAULT_ALLOWED_POLICY.messages },
-    system: { ...DEFAULT_ALLOWED_POLICY.system },
-    payments: { ...DEFAULT_ALLOWED_POLICY.payments },
-    intakes: { ...DEFAULT_ALLOWED_POLICY.intakes },
-    matters: { ...DEFAULT_ALLOWED_POLICY.matters }
-  },
-  allowed: {
-    messages: { ...DEFAULT_ALLOWED_POLICY.messages },
-    system: { ...DEFAULT_ALLOWED_POLICY.system },
-    payments: { ...DEFAULT_ALLOWED_POLICY.payments },
-    intakes: { ...DEFAULT_ALLOWED_POLICY.intakes },
-    matters: { ...DEFAULT_ALLOWED_POLICY.matters }
-  }
-};
-
-function normalizePolicyChannel(
-  raw: unknown,
-  fallback: NotificationPolicyChannel
-): NotificationPolicyChannel {
-  if (!raw || typeof raw !== 'object' || Array.isArray(raw)) {
-    return fallback;
-  }
-  const record = raw as Record<string, unknown>;
-  return {
-    push: typeof record.push === 'boolean' ? record.push : fallback.push,
-    email: typeof record.email === 'boolean' ? record.email : fallback.email
-  };
-}
-
-function normalizeNotificationPolicy(raw: unknown): NotificationPolicy {
-  if (!raw || typeof raw !== 'object' || Array.isArray(raw)) {
-    return DEFAULT_NOTIFICATION_POLICY;
-  }
-  const record = raw as Record<string, unknown>;
-  const defaultsRaw = record.defaults as Record<string, unknown> | undefined;
-  const allowedRaw = record.allowed as Record<string, unknown> | undefined;
-
-  const defaults: NotificationPolicy['defaults'] = {
-    messages: normalizePolicyChannel(defaultsRaw?.messages, DEFAULT_NOTIFICATION_POLICY.defaults.messages),
-    system: normalizePolicyChannel(defaultsRaw?.system, DEFAULT_NOTIFICATION_POLICY.defaults.system),
-    payments: normalizePolicyChannel(defaultsRaw?.payments, DEFAULT_NOTIFICATION_POLICY.defaults.payments),
-    intakes: normalizePolicyChannel(defaultsRaw?.intakes, DEFAULT_NOTIFICATION_POLICY.defaults.intakes),
-    matters: normalizePolicyChannel(defaultsRaw?.matters, DEFAULT_NOTIFICATION_POLICY.defaults.matters)
-  };
-
-  const allowed: NotificationPolicy['allowed'] = {
-    messages: normalizePolicyChannel(allowedRaw?.messages, DEFAULT_NOTIFICATION_POLICY.allowed.messages),
-    system: normalizePolicyChannel(allowedRaw?.system, DEFAULT_NOTIFICATION_POLICY.allowed.system),
-    payments: normalizePolicyChannel(allowedRaw?.payments, DEFAULT_NOTIFICATION_POLICY.allowed.payments),
-    intakes: normalizePolicyChannel(allowedRaw?.intakes, DEFAULT_NOTIFICATION_POLICY.allowed.intakes),
-    matters: normalizePolicyChannel(allowedRaw?.matters, DEFAULT_NOTIFICATION_POLICY.allowed.matters)
-  };
-
-  defaults.system = { push: true, email: true };
-  allowed.system = { push: true, email: true };
-
-  return { defaults, allowed };
-}
 
 function resolveChannelPreference(
   prefs: RemoteNotificationPreferences | null | undefined,
